+++ CHANGELOG.md
@@ -21,6 +21,7 @@ For a full diff see [`c0c63bb...master`][c0c63bb...master].
 * Renamed `Ergebnis\Environment\Production` to `Ergebnis\Environment\SystemVariables` ([#6]), by [@localheinz]
 * Renamed `Ergebnis\Environment\Test` to `Ergebnis\Environment\TestVariables` ([#9]), by [@localheinz]
 * Started throwing `Ergebnis\Environment\Exception\CouldNotSet` when a system environment variable could not be set ([#14]), by [@localheinz]
+* Started throwing `Ergebnis\Environment\Exception\CouldNotUnset` when a system environment variable could not be unset ([#15]), by [@localheinz]
 
 [c0c63bb...master]: https://github.com/ergebnis/environment-variables/compare/c0c63bb...master
 
@@ -32,5 +33,6 @@ For a full diff see [`c0c63bb...master`][c0c63bb...master].
 [#8]: https://github.com/ergebnis/environment-variables/pull/8
 [#9]: https://github.com/ergebnis/environment-variables/pull/9
 [#14]: https://github.com/ergebnis/environment-variables/pull/14
+[#15]: https://github.com/ergebnis/environment-variables/pull/15
 
 [@localheinz]: https://github.com/localheinz